fix: move e2b build scripts to turbo directory
Move build scripts from e2b/ to turbo/ directory to fix module
resolution issues. Scripts now run from turbo context where
dependencies are installed.

Changes:
- Move build.prod.ts and build.dev.ts to turbo/ directory
- Update imports to reference ../e2b/template
- Update package.json scripts to reference local build files
- Fixes "Cannot find module 'e2b'" error in CI
