+++ turbo/package.json
@@ -16,8 +16,8 @@
     "knip:production": "knip --production --strict",
     "generate-certs": "bash ../scripts/generate-certs.sh",
     "check-certs": "node packages/proxy/scripts/check-certs.js",
-    "e2b:build": "dotenv -e apps/web/.env.local -- npx tsx ../e2b/build.prod.ts",
-    "e2b:build:dev": "dotenv -e apps/web/.env.local -- npx tsx ../e2b/build.dev.ts"
+    "e2b:build": "dotenv -e apps/web/.env.local -- npx tsx build.prod.ts",
+    "e2b:build:dev": "dotenv -e apps/web/.env.local -- npx tsx build.dev.ts"
   },
   "devDependencies": {
     "@vitest/coverage-v8": "3.2.4",
